+++ make/config/jogl/gl-impl-CustomCCode-gles1.c
@@ -16,10 +16,10 @@ Java_com_jogamp_opengl_impl_es1_GLES1Impl_dispatch_1glMapBuffer(JNIEnv *env, job
 
 /*   Java->C glue code:
  *   Java package: com.jogamp.opengl.impl.es1.GLES1Impl
- *    Java method: ByteBuffer newDirectByteBuffer(long addr, int capacity);
- *     C function: jobject newDirectByteBuffer(jlong addr, jint capacity);
+ *    Java method: ByteBuffer newDirectByteBuffer(long addr, long capacity);
+ *     C function: jobject newDirectByteBuffer(jlong addr, jlong capacity);
  */
 JNIEXPORT jobject JNICALL
-Java_com_jogamp_opengl_impl_es1_GLES1Impl_newDirectByteBuffer(JNIEnv *env, jobject _unused, jlong addr, jint capacity) {
+Java_com_jogamp_opengl_impl_es1_GLES1Impl_newDirectByteBuffer(JNIEnv *env, jobject _unused, jlong addr, jlong capacity) {
   return (*env)->NewDirectByteBuffer(env, (void*) (intptr_t) addr, capacity);
 }
